#4 Select Button Done in the Main Menu.
In Menu_Main.xaml.cs we create a new EventHandler "StartClicked". Then in Window_Menu.xaml.cs we listen for that event in the method showMenu_Main().

StartClicked returns an unfinished Exercise. New boolean finished added to Exercise and a method in Routine that return the unfinished Exercises. Then in Menu_Main we return the first UnfinishedExercise.